refactor(library/math): moves "*/aggregator" to "*/aggregators"
