+++ src/library/math/aggregators/Aggregator.ts
