+++ src/library/math/aggregators/arithmeticMeanOf.ts
