+++ src/library/math/aggregators/sumOf.ts
